fix old problems

+++ database/migrations/2024_11_22_132756_create_coupons_table.php
@@ -14,9 +14,9 @@ class CreateCouponsTable extends Migration
     public function up()
     {
         Schema::create('coupons', function (Blueprint $table) {
-            $table->string('code')->primary();
-            $table->integer('pourcentage');
-            $table->timestamps();
+            $table->string('code')->primary(); // 'code' as primary key
+            $table->decimal('pourcentage', 5, 2); // 'pourcentage' as decimal
+            $table->timestamps(); // Adds 'created_at' and 'updated_at' columns
         });
     }
